requirements.txt
pyTelegramBotAPI>=4.29.1
python-dotenv>=1.0.1
SQLAlchemy>=2.0.30
asyncpg>=0.29.0
aiohttp>=3.10.11
aiosqlite>=0.20.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
